refactor(minibf): split governance into a directory module
routes/governance.rs becomes routes/governance/mod.rs and the ten
helpers that reproduce db-sync's governance_description JSON move to
routes/governance/mapping.rs with their two unit tests, mirroring the
epochs module layout. Pure move, no behavior change.
